+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   The `cover` operator is used to provide an alternative value for an expression in case of an error during evaluation.
 - New built-in function `eval()`.  
   Same as `script()`, but accepts a string instead of a file path.
+- New built-in function `replace()`.  
+  Replace strings in strings.
 - Interactive mode directives.
   - `#clear` clears the console.
   - `#header` displays the header.